refactor: 関数の共通化  - DB移行を考慮し、localStorage I/Oを共通化  - format系も共通化
