+++ utils/format.ts
@@ -1,0 +1,15 @@
+export const formatDate = (dateString: string) => {
+    const date = new Date(dateString)
+    return date.toLocaleDateString("ja-JP", {
+        year: "numeric",
+        month: "short",
+        day: "numeric",
+    })
+}
+
+export const formatAmount = (amount: number) => {
+    return new Intl.NumberFormat("ja-JP", {
+        style: "currency",
+        currency: "JPY",
+    }).format(amount)
+}
